services/backend/app/requirements/base.txt
loguru
fastapi[all]
uvicorn[standard]
watchfiles==0.21.0
asyncpg
psycopg2-binary
psycopg[binary, pool]
sqlalchemy[asyncio]
sqlalchemy_utils
